requirements.txt
fastapi[standard]==0.116.1
uvicorn
gunicorn
sqlalchemy
pydantic
requests
python-dotenv
pandas
sqlite3-0611
psycopg2-binary